separate transaction_recorder.rs
